Fix: Converti Long in Int per cachedPages.size()
- cachedPages.size() restituisce Long ma size deve restituire Int
- Aggiunto .toInt() per convertire il tipo

+++ app/src/main/java/com/github/damontecres/wholphin/util/ApiRequestPager.kt
@@ -132,7 +132,7 @@ class ApiRequestPager<T>(
                 return totalCount
             }
             // Calcola un valore approssimativo basato sulle pagine già caricate
-            val loadedPages = cachedPages.size()
+            val loadedPages = cachedPages.size().toInt()
             if (loadedPages > 0) {
                 // Usa il numero di pagine caricate + 1 per permettere il caricamento di più pagine
                 return (loadedPages + 1) * pageSize
